fix:move software credential container into main
